+++ packages/web3-eth/src/utils/prepare_transaction_for_signing.ts
@@ -126,7 +126,6 @@ export const prepareTransactionForSigning = async (
 	privateKey?: HexString | Uint8Array,
 	fillGasPrice = false,
 ) => {
-	console.log("preparetransaction for signing")
 	const populatedTransaction = (await transactionBuilder({
 		transaction,
 		web3Context,