+++ packages/web3-eth/src/validation.ts
@@ -250,13 +250,10 @@ export const validateFeeMarketGas = (transaction: InternalTransaction) => {
  * legacy gas (type 0x0 and 0x1) OR fee market transactions (0x2)
  */
 export const validateGas = (transaction: InternalTransaction) => {
-	console.log("validateGas")
-	console.log(transaction)
 	const gasPresent = !isNullish(transaction.gas) || !isNullish(transaction.gasLimit);
 	const legacyGasPresent = gasPresent && !isNullish(transaction.gasPrice);
 	const feeMarketGasPresent =
-		!isNullish(transaction.maxPriorityFeePerGas) &&
-		!isNullish(transaction.maxFeePerGas);
+		!isNullish(transaction.maxPriorityFeePerGas) && !isNullish(transaction.maxFeePerGas);
 
 	if (!legacyGasPresent && !feeMarketGasPresent)
 		throw new MissingGasError({