+++ packages/web3-eth/test/unit/prepare_transaction_for_signing.test.ts
@@ -53,7 +53,7 @@ describe('prepareTransactionForSigning', () => {
 			) => {
 				// (i.e. requestManager, blockNumber, hydrated params), but that doesn't matter for the test
 				jest.spyOn(ethRpcMethods, 'estimateGas').mockImplementation(
-					// @ts-expect-error - Mocked implementation doesn't have correct method signaturea
+					// @ts-expect-error - Mocked implementation doesn't have correct method signature
 					() => expectedTransaction.gas,
 				);
 				// @ts-expect-error - Mocked implementation doesn't have correct method signature